# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Rare District — Nigerian fashion marketplace API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /products:
    get:
      operationId: listProducts
      tags: [Products]
      summary: List products with optional filters
      parameters:
        - name: category
          in: query
          schema:
            type: string
        - name: vendorId
          in: query
          schema:
            type: integer
        - name: minPrice
          in: query
          schema:
            type: number
        - name: maxPrice
          in: query
          schema:
            type: number
        - name: size
          in: query
          schema:
            type: string
        - name: search
          in: query
          schema:
            type: string
        - name: sortBy
          in: query
          schema:
            type: string
            enum: [newest, price_asc, price_desc, popular]
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 24
      responses:
        "200":
          description: Paginated product list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProductPage"

    post:
      operationId: createProduct
      tags: [Products]
      summary: Create a new product listing (vendor only)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProductInput"
      responses:
        "201":
          description: Created product
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
        "403":
          description: Vendor not approved
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
  /products/featured:
    get:
      operationId: listFeaturedProducts
      tags: [Products]
      summary: Get featured / editorial products
      parameters:
        - name: limit
          in: query
          schema:
            type: integer
            default: 8
      responses:
        "200":
          description: Featured products
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Product"
  /products/trending:
    get:
      operationId: listTrendingProducts
      tags: [Products]
      summary: Get trending products (most added to wardrobes)
      parameters:
        - name: limit
          in: query
          schema:
            type: integer
            default: 8
      responses:
        "200":
          description: Trending products
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Product"
  /products/{id}:
    get:
      operationId: getProduct
      tags: [Products]
      summary: Get a product by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Product
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"

    patch:
      operationId: updateProduct
      tags: [Products]
      summary: Update a product (vendor or admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProductUpdate"
      responses:
        "200":
          description: Updated product
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"

    delete:
      operationId: deleteProduct
      tags: [Products]
      summary: Delete a product (vendor or admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # ── CATEGORIES ────────────────────────────────────────────────────────────
...
  /referrals/share:
    post:
      operationId: generateShareLink
      tags: [Referrals]
      summary: Generate a trackable share link for a product
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ShareLinkInput"
      responses:
        "200":
          description: Share link
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ShareLink"

  # ── REVIEWS ───────────────────────────────────────────────────────────────
  /products/{id}/reviews:
    get:
      operationId: getProductReviews
      tags: [Reviews]
      summary: List reviews for a product
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Reviews
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Review"

    post:
      operationId: createReview
      tags: [Reviews]
      summary: Create a review for a product
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ReviewInput"
      responses:
        "201":
          description: Created review
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Review"

  # ── STOREFRONT ────────────────────────────────────────────────────────────
...
components:
  schemas:
...
    ErrorEnvelope:
      type: object
      required: [error]
      properties:
        error:
          type: string
...
    User:
      type: object
      required: [id, email, role, referralCode, createdAt]
      properties:
        id:
          type: integer
        email:
          type: string
        name:
          type: ["string", "null"]
        avatarUrl:
          type: ["string", "null"]
        role:
          type: string
          enum: [shopper, vendor, admin]
        referralCode:
          type: string
        referredBy:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
...
    Vendor:
      type: object
      required: [id, userId, brandName, status, payoutBalance, createdAt]
      properties:
        id:
          type: integer
        userId:
          type: integer
        brandName:
          type: string
        description:
          type: ["string", "null"]
        logoUrl:
          type: ["string", "null"]
        website:
          type: ["string", "null"]
        bankName:
          type: ["string", "null"]
        accountNumber:
          type: ["string", "null"]
        accountName:
          type: ["string", "null"]
        status:
          type: string
          enum: [pending, approved, rejected]
        commissionRateOverride:
          type: ["number", "null"]
        payoutBalance:
          type: number
        adminNote:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
        user:
          $ref: "#/components/schemas/User"
...
    Product:
      type: object
      required: [id, vendorId, name, price, currency, stock, isActive, createdAt]
      properties:
        id:
          type: integer
        vendorId:
          type: integer
        name:
          type: string
        description:
          type: ["string", "null"]
        price:
          type: number
        currency:
          type: string
          default: NGN
        category:
          type: ["string", "null"]
        sizes:
          type: array
          items:
            type: string
        images:
          type: array
          items:
            type: string
        stock:
          type: integer
        isActive:
          type: boolean
        isFeatured:
          type: boolean
        wardrobeCount:
          type: integer
        averageRating:
          type: ["number", "null"]
        reviewCount:
          type: integer
        createdAt:
          type: string
          format: date-time
        vendor:
          $ref: "#/components/schemas/Vendor"
    ProductInput:
      type: object
      required: [name, price, stock]
      properties:
        name:
          type: string
          minLength: 2
        description:
          type: string
        price:
          type: number
          minimum: 0
        category:
          type: string
        sizes:
          type: array
          items:
            type: string
        images:
          type: array
          items:
            type: string
        stock:
          type: integer
          minimum: 0
        isFeatured:
          type: boolean
    ProductUpdate:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
        price:
          type: number
        category:
          type: string
        sizes:
          type: array
          items:
            type: string
        images:
          type: array
          items:
            type: string
        stock:
          type: integer
        isActive:
          type: boolean
        isFeatured:
          type: boolean
    ProductPage:
      type: object
      required: [items, total, page, limit]
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/Product"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer

    # ── Category ─────────────────────────────────────────────────────────────
...
    ShareLinkInput:
      type: object
      required: [productId]
      properties:
        productId:
          type: integer
    ShareLink:
      type: object
      required: [url, shareCode]
      properties:
        url:
          type: string
        shareCode:
          type: string

    # ── Reviews ───────────────────────────────────────────────────────────────
    Review:
      type: object
      required: [id, productId, userId, rating, createdAt]
      properties:
        id:
          type: integer
        productId:
          type: integer
        userId:
          type: integer
        rating:
          type: integer
          minimum: 1
          maximum: 5
        comment:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
        user:
          $ref: "#/components/schemas/User"
    ReviewInput:
      type: object
      required: [rating]
      properties:
        rating:
          type: integer
          minimum: 1
          maximum: 5
        comment:
          type: string

    # ── Storefront ────────────────────────────────────────────────────────────
...