# lib/api-zod/src/generated/types/productInput.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * Api
 * Rare District — Nigerian fashion marketplace API
 * OpenAPI spec version: 0.1.0
 */

export interface ProductInput {
  /** @minLength 2 */
  name: string;
  description?: string;
  /** @minimum 0 */
  price: number;
  category?: string;
  sizes?: string[];
  images?: string[];
  /** @minimum 0 */
  stock: number;
  isFeatured?: boolean;
}